+++ Makefile
@@ -219,7 +219,6 @@ SOURCE_FILES = \
   BoundsInference.cpp \
   BranchVisitors.cpp \
   Buffer.cpp \
-  CodeGen.cpp \
   CodeGen_ARM.cpp \
   CodeGen_C.cpp \
   CodeGen_GPU_Dev.cpp \
@@ -316,12 +315,10 @@ HEADER_FILES = \
   BranchVisitors.h \
   Buffer.h \
   buffer_t.h \
-  CodeGen.h \
   CodeGen_ARM.h \
   CodeGen_C.h \
   CodeGen_GPU_Dev.h \
   CodeGen_GPU_Host.h \
-  CodeGen.h \
   CodeGen_LLVM.h \
   CodeGen_MIPS.h \
   CodeGen_OpenCL_Dev.h \
